+++ intelmq/tests/bots/collectors/tcp/test_collector.py
@@ -1,184 +1,179 @@
-# -*- coding: utf-8 -*-
+# uncompyle6 version 3.1.3
+# Python bytecode 3.6 (3379)
+# Decompiled from: Python 3.6.3 (default, Oct  3 2017, 21:45:48) 
+# [GCC 7.2.0]
+# Embedded file name: /var/www/intelmq-repository/intelmq/tests/bots/collectors/tcp/test_collector.py
+# Compiled at: 2018-03-28 21:30:52
+# Size of source mod 2**32: 6074 bytes
 """
-Testing HTTP collector
+Testing TCP collector
 """
-
 import socket
+import struct
 import threading
 import unittest
+from multiprocessing import Process
 from time import sleep
 
 import intelmq.lib.test as test
 import intelmq.lib.utils as utils
 from intelmq.bots.collectors.tcp.collector import TCPCollectorBot
+from intelmq.bots.outputs.tcp.output import TCPOutputBot
 from intelmq.lib.message import Event
 from intelmq.lib.message import MessageFactory
 from intelmq.lib.utils import base64_decode
 
-OUTPUT = [{"__type": "Report",
-           "feed.name": "Example feed",
-           "feed.accuracy": 100.,
-           "feed.url": "http://localhost/two_files.tar.gz",
-           "raw": utils.base64_encode('bar text\n'),
-           },
-          {"__type": "Report",
-           "feed.name": "Example feed",
-           "feed.accuracy": 100.,
-           "feed.url": "http://localhost/two_files.tar.gz",
-           "raw": utils.base64_encode('foo text\n'),
-           },
-          ]
-
-USECASE1 = {"classification.taxonomy": "malicious code", "classification.type": "c&c", "feed.accuracy": 50.0,
-            "feed.name": "abuse-ch-zeus-tracker-domains",
-            "feed.url": "https://zeustracker.abuse.ch/blocklist.php?download=baddomains",
-            "malware.name": "zeus", "raw": "bHplZGlhbWlrZS50cmFkZQ==", "source.fqdn": "lzediamike.trade",
-            "time.observation": "2018-03-20T14:05:03+00:00"}
-
-INPUT1 = {"classification.taxonomy": "malicious code",
-          "classification.type": "c&c",
-          "feed.name": "Example feed",
-          "feed.accuracy": 100.,
-          "feed.url": "http://localhost/two_files.tar.gz",
-          "raw": utils.base64_encode('bar text\n'),
-          "time.observation": "2018-03-20T14:05:03+00:00"
-          }
+PORT = 5011
+SEPARATOR = '\n'
+INPUT1 = {'classification.taxonomy': 'malicious code',
+          'classification.type': 'c&c',
+          'feed.name': 'Example feed',
+          'feed.accuracy': 100.0,
+          'feed.url': 'http://localhost/two_files.tar.gz',
+          'raw': utils.base64_encode('bar text\n'),
+          'time.observation': '2018-03-20T14:05:03+00:00'}
 REPORT1 = {
     'raw': 'eyJjbGFzc2lmaWNhdGlvbi50YXhvbm9teSI6ICJtYWxpY2lvdXMgY29kZSIsICJjbGFzc2lmaWNhdGlvbi50eXBlIjogImMmYyIsICJmZWVkLm5hbWUiOiAiRXhhbXBsZSBmZWVkIiwgImZlZWQuYWNjdXJhY3kiOiAxMDAuMCwgImZlZWQudXJsIjogImh0dHA6Ly9sb2NhbGhvc3QvdHdvX2ZpbGVzLnRhci5neiIsICJyYXciOiAiWW1GeUlIUmxlSFFLIiwgInRpbWUub2JzZXJ2YXRpb24iOiAiMjAxOC0wMy0yMFQxNDowNTowMyswMDowMCJ9',
-    'feed.name': 'Example feed', 'feed.accuracy': 100.0}
+    'feed.name': 'Example feed',
+    'feed.accuracy': 100.0}
+INPUT2 = {'feed.name': 'Example feed 2',
+          'feed.accuracy': 100.0,
+          'feed.url': 'http://localhost/two_files.tar.gz',
+          'raw': utils.base64_encode('foo text\n')}
+ORIGINAL_DATA = ('some random input{}another line').format(SEPARATOR)
 
-INPUT2 = {
-    "feed.name": "Example feed",
-    "feed.accuracy": 100.,
-    "feed.url": "http://localhost/two_files.tar.gz",
-    "raw": utils.base64_encode('foo text\n'),
-}
 
-
-@test.skip_local_web()
-class TestIntelMQCollectorBot(test.BotTestCase, unittest.TestCase):
-    """
-    A TestCase for HTTPCollectorBot.
-    """
-
-    @classmethod
-    def set_bot(cls):
-        cls.PORT = 5011
-        cls.bot_reference = TCPCollectorBot
-        cls.sysconfig = {'http_url': 'http://localhost/two_files.tar.gz',
-                         'extract_files': True,
-                         'feed': 'Example feed',
-                         'ip': "localhost",
-                         "port": cls.PORT,
-                         "separator": "\n"
-                         }
-        cls.separator = b"\n"
+class Client:
+    """ You find here an example of a non-intelmq client that might connect to the bot. """
 
     def _client(self):
         sleep(1)
         connection = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        connection.connect(("localhost", self.PORT))
+        connection.connect(('localhost', PORT))
         return connection
 
-    def _test_collecting_single_message(self):
-        """ Test if correct Events have been produced. """
+    def random_client(self):
+        connection = self._client()
+        d = bytes(ORIGINAL_DATA.split(SEPARATOR)[0], 'UTF-8')
+        msg = struct.pack('>I', len(d)) + d
+        connection.sendall(msg)
+        d = bytes(ORIGINAL_DATA.split(SEPARATOR)[1], 'UTF-8')
+        msg = struct.pack('>I', len(d)) + d
+        connection.sendall(msg)
+        connection.close()
 
-        def client():
-            data = Event(INPUT1).to_json(hierarchical=False)
-            text = utils.encode(data) + self.separator
 
-            connection = self._client()
-            connection.sendall(text)
-            connection.close()
+class TestTCPOutputBot(test.BotTestCase, unittest.TestCase):
+    """ Instance of TCPOutput bot might help to simulate a real world situation with reconnecting etc. """
 
-        thread = threading.Thread(target=client)
+    @classmethod
+    def set_bot(cls):
+        cls.bot_reference = TCPOutputBot
+        cls.sysconfig = {'hierarchical_output': False,
+                         'ip': 'localhost',
+                         'port': PORT
+                         }
+
+    def _delayed_start(self):
+        sleep(1)
+        self.assertEqual = lambda *args, **kwargs: True
+        self.run_bot(iterations=len(self.input_message))
+
+
+class TestTCPCollectorBot(test.BotTestCase, unittest.TestCase):
+    """
+    A TestCase for TCPCollectorBot.
+    """
+
+    @classmethod
+    def set_bot(cls):
+        cls.bot_reference = TCPCollectorBot
+        cls.sysconfig = {'http_url': 'http://localhost/two_files.tar.gz', 'extract_files': True,
+                         'feed': 'Example feed',
+                         'ip': 'localhost',
+                         'port': PORT
+                         }
+
+    def test_random_input(self):
+        """ Check how we handle a random input, coming from an unknown source. We should put all the data to report['raw']. """
+        thread = threading.Thread(target=Client().random_client)
         thread.start()
         self.input_message = None
         self.run_bot()
+        self.assertOutputQueueLen(2)
+        generated_report = MessageFactory.unserialize(self.get_output_queue()[1], harmonization=self.harmonization,
+                                                      default_type='Event')
+        self.assertEqual(base64_decode(generated_report['raw']), ORIGINAL_DATA.split(SEPARATOR)[1])
 
-        self.assertOutputQueueLen(1)
-        line = self.get_output_queue()[0]
-        generated_report = MessageFactory.unserialize(line, harmonization=self.harmonization, default_type='Event')
-        original_message = MessageFactory.unserialize(base64_decode(generated_report["raw"]), harmonization=self.harmonization,
-                                               default_type='Event')
-        self.assertDictEqual(original_message, INPUT1)
-        del generated_report["time.observation"]
-        self.assertDictEqual(generated_report, REPORT1)
+    def test_intelmq_exchange(self):
+        """ Test if correct Events have been produced, sent from a TCP Output of another IntelMQ instance.
+        We spawn independent process of the TCPOutput bot that sends a bunch of messages.
+        """
+        bot = TestTCPOutputBot()
+        bot.setUpClass()
+        bot.input_message = []
+        msg_count = 100 # 100
+        for i in range(msg_count):
+            bot.input_message.append(Event(INPUT1))
+        (Process(target=bot._delayed_start)).start()
+        self.run_bot()
+        self.assertOutputQueueLen(msg_count)
+        for msg in self.get_output_queue():
+            report = MessageFactory.unserialize(msg, harmonization=self.harmonization, default_type='Event')
+            del report['time.observation']
+            self.assertDictEqual(report, REPORT1)
 
-    def _test_collecting_chunked_messages(self):
-        """ Divide a message into chunks per n characters and send it 4 times consecutively. """
+    def test_chunked_msg(self):
+        """ Test if correct Events have been produced, sent from a TCP Output of another IntelMQ instance,
+        when we divide the stream into chunks per n characters and send it 4 times consecutively.
+        That way we simulate network throttling. """
 
-        def client():
-            connection = self._client()
-            data = Event(INPUT2).to_json(hierarchical=False)
-            stream = utils.encode(data) + self.separator
+        def chunked_process_replacement(self):
+            event = self.receive_message()
+            data = event.to_json(hierarchical=self.parameters.hierarchical_output)
+            d = utils.encode(data)
+            msg = struct.pack('>I', len(d)) + d
+            chunk_length = 40
+            for chunk in [msg[i:i + chunk_length] for i in range(0, len(msg), chunk_length)]:
+                self.con.sendall(chunk)
 
-            n = 3
-            for i in range(4):
-                for chunk in [stream[i:i + n] for i in range(0, len(stream), n)]:
-                    connection.sendall(chunk)
-            connection.close()
+        TCPOutputBot._process = TCPOutputBot.process
+        TCPOutputBot.process = chunked_process_replacement
+        self.test_intelmq_exchange()
+        TCPOutputBot.process = TCPOutputBot._process
 
-        thread = threading.Thread(target=client)
+    def test_multiple_bots(self):
+        """ Let's simulate multiple IntelMQ instances are pushing the events at once!
+            Every message must be queued.
+            Note that if too much clients want connect, connections are refused.
+        """
+        client_count = 5
+        msg_count = 300
+        for _ in range(client_count):
+            bot = TestTCPOutputBot()
+            bot.setUpClass()
+            bot.input_message = []
+            for i in range(msg_count):
+                bot.input_message.append(Event(INPUT1))
+            Process(target=bot._delayed_start).start()
+
+        thread = threading.Thread(target=Client().random_client)
         thread.start()
         self.input_message = None
-        self.run_bot()
 
-        self.assertOutputQueueLen(4)
-        line = self.get_output_queue()[2]
-        generated_report = MessageFactory.unserialize(line, harmonization=self.harmonization, default_type='Event')
-        original_message = MessageFactory.unserialize(base64_decode(generated_report["raw"]), harmonization=self.harmonization,
-                                               default_type='Event')
-        self.assertDictEqual(original_message, INPUT2)
+        # can't use standard .bot_run(iteration) or .start() because shutdown() would be called
+        # and we need to handle multiple connections
+        self.prepare_bot()
+        self.bot._Bot__source_pipeline = self.pipe
+        self.bot._Bot__destination_pipeline = self.pipe
+        for _ in range(client_count+1):
+            # every single calling of process() method will serve to a single connection
+            self.bot.process()
+        self.bot.stop() # let's call shutdown() and free up binded address
 
-    def test_invalid_input(self):
-        """ Check how we handle an invalid input. """
-
-        def client():
-            connection = self._client()
-            data = Event(INPUT2).to_json(hierarchical=False)
-            stream = b"some invalid input" + self.separator
-            connection.sendall(stream)
-            connection.close()
-
-        thread = threading.Thread(target=client)
-        thread.start()
-        self.input_message = None
-        self.run_bot()
-
-        import ipdb; ipdb.set_trace()
-
-        self.assertOutputQueueLen(4)
-        line = self.get_output_queue()[2]
-        generated_report = MessageFactory.unserialize(line, harmonization=self.harmonization, default_type='Event')
-        original_message = MessageFactory.unserialize(base64_decode(generated_report["raw"]), harmonization=self.harmonization,
-                                               default_type='Event')
-        self.assertDictEqual(original_message, INPUT2)
+        self.assertOutputQueueLen(client_count * msg_count + 2)
 
 
-if __name__ == '__main__':  # pragma: no cover
+if __name__ == '__main__':
     unittest.main()
-
-"""
-XXXtemporary debug notes: 
-m = '{"raw": "YmVzdGRvdmUuaW4udWE=", "feed.url": "https://zeustracker.abuse.ch/blocklist.php?download=baddomains", "classification.type": "c&c", "time.observation": "2018-03-20T14:05:03+00:00", "feed.accuracy": 50.0, "feed.name": "abuse-ch-zeus-tracker-domains", "source.fqdn": "bestdove.in.ua", "malware.name": "zeus", "classification.taxonomy": "malicious code"}'
-data = b'{"classification.taxonomy": "malicious code", "time.observation": "2018-03-20T14:05:03+00:00", "feed.name": "abuse-ch-zeus-tracker-domains", "classification.type": "c&c", "feed.accuracy": 50.0, "source.fqdn": "blogerjijer.pw", "malware.name": "zeus", "feed.url": "https://zeustracker.abuse.ch/blocklist.php?download=baddomains", "raw": "YmxvZ2VyamlqZXIucHc="}|'
-events = []
-for m in data.split(self.separator):
-    if m:
-        event = Event(Event.unserialize(utils.decode(m)))
-        import traceback;
-        import ipdb; ipdb.set_trace()
-        self.send_message(event)
-        super(CollectorBot, self).send_message(event)
-
-debug to send -> use in the tests to send a chunked-messaged
-s = "" "{"classification.taxonomy": "malicious code","classification.type": "c&c","feed.accuracy": 50.0,"feed.name": "abuse-ch-zeus-tracker-domains","feed.url": "https://zeustracker.abuse.ch/blocklist.php?download=baddomains","malware.name": "zeus","raw": "bHplZGlhbWlrZS50cmFkZQ==","source.fqdn": "lzediamike.trade","time.observation": "2018-03-20T14:05:03+00:00"}"" "
-event = Event(Event.unserialize(s))
-data = event.to_json(hierarchical=self.parameters.hierarchical_output)
-l = utils.encode(data) + self.separator
-n = 3
-for chunk in [l[i:i + n] for i in range(0, len(l), n)]:
-    print(chunk)
-    self.con.sendall(chunk)
-"""
+# okay decompiling /tmp/ram/test_collector.cpython-36.pyc